+++ src/main/java/fr/HtSTeam/HtS/Options/Options/LootTables/CrateOption.java
@@ -8,52 +8,44 @@ import java.nio.file.Paths;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 
-import fr.HtSTeam.HtS.Options.OptionRegister;
-import fr.HtSTeam.HtS.Options.Structure.Alterable;
+import fr.HtSTeam.HtS.Options.GUIRegister;
 import fr.HtSTeam.HtS.Options.Structure.OptionBuilder;
-import fr.HtSTeam.HtS.Utils.FileExtractor;
+import fr.HtSTeam.HtS.Utils.Files.FileExtractor;
 
-public class CrateOption extends OptionBuilder implements Alterable {
-	
-	private boolean activate = false;
-	
+public class CrateOption extends OptionBuilder<Boolean>{
+		
 	public CrateOption() {
-		super(Material.BROWN_SHULKER_BOX, "Crate", "§4Désactivé", "Désactivé", OptionRegister.loottables);
+		super(Material.BROWN_SHULKER_BOX, "Crate", "§4Désactivé", false, GUIRegister.loottables, false);
 	}
 
 	@Override
 	public void event(Player p) {
-		activate =! activate;
-		setState(activate);
+		setState(!getValue());
 	}
 
 	@Override
-	public void setState(boolean value) {
-		activate = value;
-		if(value) {
+	public void setState(Boolean value) {
+		if(value && !getValue()) {
 			try {
 				FileExtractor.extractFile(FileExtractor.lt + "crate.json", FileExtractor.wdir + FileExtractor.Cdir);
-				setValue("Activé");
 				getItemStack().setLore("§2Activé");
 			} catch (IOException | URISyntaxException e) {
 				e.printStackTrace();
 			}
-		} else {
+		} else if(!value && getValue()){
 			try {
 				Files.delete(Paths.get(FileExtractor.wdir + FileExtractor.Cdir + "crate.json"));
-				setValue("Désactivé");
 				getItemStack().setLore("§4Désactivé");
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
 		}
+		setValue(value);
 		parent.update(this);
 	}
-	
-	public boolean isActivated() {
-		if (getValue().equals("Activé"))
-			return true;
-		else
-			return false;
+
+	@Override
+	public String description() {
+		return null;
 	}
 }
