+++ src/main/java/fr/HtSTeam/HtS/Options/Options/LootTables/GhastOption.java
@@ -8,39 +8,32 @@ import java.nio.file.Paths;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Player;
 
-import fr.HtSTeam.HtS.Options.OptionRegister;
-import fr.HtSTeam.HtS.Options.Structure.Alterable;
+import fr.HtSTeam.HtS.Options.GUIRegister;
 import fr.HtSTeam.HtS.Options.Structure.OptionBuilder;
-import fr.HtSTeam.HtS.Utils.FileExtractor;
 import fr.HtSTeam.HtS.Utils.ItemStackBuilder;
+import fr.HtSTeam.HtS.Utils.Files.FileExtractor;
 
-public class GhastOption extends OptionBuilder implements Alterable {
-
-	private boolean activate = false;
+public class GhastOption extends OptionBuilder<Boolean>{
 
 	public GhastOption() {
-		super(new ItemStackBuilder(EntityType.GHAST, 1, "§rGhast", "§4Désactivé"), "Désactivé", OptionRegister.loottables);
+		super(new ItemStackBuilder(EntityType.GHAST, 1, "§rGhast", "§4Désactivé"), false, GUIRegister.loottables, false);
 	}
 
 	@Override
 	public void event(Player p) {
-		activate  =! activate;
-		setState(activate);
+		setState(!getValue());
 	}
 
 	@Override
-	public void setState(boolean value) {
-		activate = value;
-		if(value) {
+	public void setState(Boolean value) {
+		if(value && !getValue()) {
 			try {
 				FileExtractor.extractFile(FileExtractor.lt + "ghast.json", FileExtractor.wdir + FileExtractor.Edir);
-				setValue("Activé");
 				getItemStack().setLore("§2Activé");
 			} catch (IOException | URISyntaxException e) {
 				e.printStackTrace();
 			}
-		} else {
-			setValue("Désactivé");
+		} else if(!value && getValue()){
 			getItemStack().setLore("§4Désactivé");
 			try {
 				Files.delete(Paths.get(FileExtractor.wdir + FileExtractor.Edir + "ghast.json"));
@@ -48,13 +41,12 @@ public class GhastOption extends OptionBuilder implements Alterable {
 				e.printStackTrace();
 			}
 		}
+		setValue(value);
 		parent.update(this);
 	}
-	
-	public boolean isActivated() {
-		if (getValue().equals("Activé"))
-			return true;
-		else
-			return false;
-	}
+
+	@Override
+	public String description() {
+		return null;
+	}	
 }
